pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "GuilhermeHalter", email = "guisam.halter@gmail.com"},
]
dependencies = [
    "django>=4.2.7",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.1",
    "setuptools>=68.2.2",
    "python-magic>=0.4.27",
    "Pillow>=10.1.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"
[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.12.0",
    "black>=23.11.0",
]

[tool.black]
line-length = 120